requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
connectorx==0.2.3
contourpy==1.2.1
cycler==0.12.1
DateTime==5.5
fonttools==4.51.0
greenlet==3.0.3
idna==3.7
joblib==1.4.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
polars==0.20.23
psycopg2-binary==2.9.9
PuLP==2.8.0
pyarrow==16.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
SQLAlchemy==2.0.29
threadpoolctl==3.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
zope.interface==6.3
